app/requirements.txt
aio-pika==6.6.1
aiohttp==3.6.2
aioredis==1.3.1
aiormq==3.2.2
amqp==2.6.0
asgiref==3.2.7
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.2
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
daphne==2.5.0
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
drf-yasg==1.17.1
future==0.18.2
h11==0.9.0
hiredis==1.0.1
httptools==0.1.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.1
incremental==17.5.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.10
MarkupSafe==1.1.1
msgpack==0.6.2
multidict==4.7.6
packaging==20.4
pamqp==2.3.0
pika==1.1.0
psycopg2==2.7.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
vine==1.3.0
websockets==8.1
yarl==1.4.2
zipp==3.1.0
zope.interface==5.1.0
